changed type of s

+++ src/DataTypes.java
@@ -3,8 +3,8 @@ import java.util.List;
 public class DataTypes {
     // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(List<Integer> numbers) {
-
-        int s = 0;
+        // changed type of s from int to `long` as long can handle larger values than int
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
